Remove special no-op rule from TipInputModeManager::OnSetFocus
With this commit, one remaining behavior difference between Mozc
and MS-IME in Windows 11 22H2 will be addressed.

Mozc has had a special no-op rule when switching input focus and the
InputScope specified in the newly focused field is the same as the one
that was previously observed in the same process.  MS-IME, on the other
hand, does not have such a special no-op rule and always updates its
input mode upon switching input focus.

For better app compatibility, let's make it compatible with MS-IME's.

Closes #826.

PiperOrigin-RevId: 574371235