Automated g4 rollback of changelist 575114935.
*** Reason for rollback ***

This rollback turned out to be conflicting, and we will retry rollback steps.

*** Original change description ***

Automated g4 rollback of changelist 568386718.

*** Reason for rollback ***

This CL in A release will cause unintended regression because corresponding CL is also rollbacked.

Because the CL is working perfectly with another CLs, this will be rollbacked again after cherry-picked into A release.

*** Original change description ***

Remove special treatment of 'サジェストのみ(SUGGESTION_ONLY)' from Android User Dictionary files.

This implementation is introduced in cl/424818781 to provid...

***

PiperOrigin-RevId: 575153063