Automated g4 rollback of changelist 575990015.
*** Reason for rollback ***

Because the rollback was for cherry-picking, I would like to rollback the rollback again.

*** Original change description ***

Automated g4 rollback of changelist 568386718.

*** Reason for rollback ***

This CL causes bug without another CL which is already rollbacked, so this CL also has to be rollbacked.
There are another trial for rollbacking, but that failed due to conflict.

*** Original change description ***

Remove special treatment of 'サジェストのみ(SUGGESTION_ONLY)' from Android User Dictionary files.

This implementation is introduced in cl/424818781 to provide special behavior for user dictionary en...

***

PiperOrigin-RevId: 576032587