Automated g4 rollback of changelist 575979626.
*** Reason for rollback ***

This rollback was for cherry-picking, and now it can be rollbacked again

*** Original change description ***

Automated g4 rollback of changelist 575072274.

*** Reason for rollback ***

This CL causes conflict when rollbacking cl/568386718. To resolve the conflict, we have to first rollback this CL.

This rollback will be rollbacked again after issues are resolved.

*** Original change description ***

Fix problem that NO_POS is not working in Android Gboard

- The cause of the bug was that while user_pos_*.data is in the dictionary side, Android Gboard did not have an updated version of dictionary dat...

***

PiperOrigin-RevId: 576057664