Enable Hardware Enforced Stack Protection in Windows
With this commit, both Mozc and Qt executables are build with the
following flag set:

  IMAGE_DLLCHARACTERISTICS_EX_CET_COMPAT [1]

This is an optional flag that tells the OS that executables are
compatible with Intel Control Flow Enforcement Technology (CET) [2].

The flag simply ignored on environment where the feature is not
available.  There must be no user observable behavior change.

Closes #835.

 [1]: https://learn.microsoft.com/en-us/windows/win32/debug/pe-format#extended-dll-characteristics
 [2]: https://www.intel.com/content/www/us/en/developer/articles/technical/technical-look-control-flow-enforcement-technology.html

PiperOrigin-RevId: 576253756