Create an import library for 'msctf.dll'
This is a preparation to remove the dependency on CoCreateInstance from
mozc_tip{32,64}.dll to comply with GUID_TFCAT_TIPCAP_COMLESS (#837).

The idea is the same as my previous commit for input.dll [1].

Note that this commit only creates an import library with unit tests,
and the production code has not been updated to use it yet.  Thus there
must be no observable behavior change (yet).

#codehealth

 [1]: f470414b26d0e0f13f2ec865e0f3a1d2046a5398

PiperOrigin-RevId: 576415439