Remove GetProcAddress to TF_CreateCategoryMgr
This is a preparation to remove the dependency on CoCreateInstance from
mozc_tip{32,64}.dll to comply with GUID_TFCAT_TIPCAP_COMLESS (#837).

With my previous commit [1], we now create an import library for
msctf.dll as part of our build process.  With this commit,
mozc_tip{32,64}.dll actually start importing TF_CreateCategoryMgr from
msctf.dll in an implicit manner.  This should be safer and more
efficient than relying on GetProcAddress as long as the API in question
is guaranteed to exist.

Note that we have already done the same optimization for input.dll [2].

There should be no user observable behavior change.

#codehealth

 [1]: 64c946116b263f726801bdcab57a5c15dc630942
 [2]: f470414b26d0e0f13f2ec865e0f3a1d2046a5398

PiperOrigin-RevId: 576674633