Fix unexpected errors when launching process in Windows
This is a follow up to our previous commit [1], after which

  WinSandbox::SpawnSandboxedProcess

stopped updating 'pid' parameter any more.  Its no longer returning
process ID does not mean that the specified process does not start, and
the client just works if the specified process is already running.

The only annoyance is that when the user first try to launch
'mozc_server' or 'mozc_renderer', the relevant code may conclude that
it's failed to spawn such process then given up to launch the process
again.  When this happens, the user needs to restart the application or
switch to another IME then switch back to Mozc.

Anyway, the above annoyance should disappear with this commit.

Closes #839.

 [1]: c1fcbbcd2b25d3970c7dc51cd5e507f174cdf31c

PiperOrigin-RevId: 576742515