Prelaunch processes from post install dialog in Windows
This is a follow up to my previous commit [1], which started executing
the following command

 mozc_broker --mode=prelaunch_processes

to pre-launch 'mozc_server' and 'mozc_renderer' whenever the user logs
in on Windows because these processes cannot be launched from sandboxed
processes such as 'SearchHost.exe'.

With this commit, effectively the same workaround will be performed
after installing Mozc and while executing post install dialog.

Closes #845.

 [1]: 048808221cae6180c7a6ae6623e13c58490f30ed

PiperOrigin-RevId: 582383447