Simplify CustomAction sequence definitions
As a preparation to simplify mozc::win32::UninstallHelper (#866), this
commit updates WiX installer script without changing the actual
behavior.

Currently CustomAction sequence around 'InstallFinalize' is not obvious
as we mix 'Before' and 'After' rules. With this commit, all the actions
before 'InstallFinalize' are constraint with 'Before' rule, which helps
us understand how these actions are ordered.

Note that 'Before' and 'After' rules are hint for WiX to order given
actions. There must be no change in the generated installer files.

#codehealth

PiperOrigin-RevId: 603068684