Switch to `macos-14` GitHub Actions runner (#868)
With this commit macOS GitHub Actions start running on 'macos-14'
runner, which now runs on Apple silicon.

One observable change is that 'Mozc_arm64.pkg' from 'build_arm64' job
is a bit smaller because we no longer need to build Qt6 as universal
binaries.

Python version is also updated as Python 3.9 is not available for
'macos-14' runner right now.

Closes #868.

PiperOrigin-RevId: 603665115