Use the default Python in `macos-14` image
This is a follow up commit to my previous commit [1], which upgraded
GitHub Actions runner for macOS builds to 'macos-14' (#868).

While we still support Python 3.9 to build Mozc on macOS environment,
currently we cannot use 'setup-python' to set up Python 3.9 in
'macos-14' runner [2]. Also we don't use 'setup-python' in other
platforms such as Windows build.

For simplicity, let's just use the default python in the runner [3],
which is Python 3.11 as of writing. Then we should no longer see
errors from 'setup-python' in the build summary [4].

#codehealth

 [1]: b8cd34e1f735cd0e2815861f1113cef648a063c2
 [2]: https://github.com/actions/setup-python/issues/808
 [3]: https://github.com/actions/virtual-environments/blob/main/images/macos/macos-14-Readme.md
 [4]: https://github.com/actions/setup-python/issues/809

PiperOrigin-RevId: 604223079