Replace `Utf8ToCodepoint` in `ImmutableConverter` with `Utf8AsChars32`
To do this, this patch changes the signature of `AddCharacterTypeBasedNodes` to take `string_view` instead of `const char*`. It also changes its caller, `Lookup`, to use `string_view` from the beginning, as `string_view` and `substr` are generally safer than a pointer and a length.

This patch should have no behavior changes.

PiperOrigin-RevId: 614365499