Add space insertion mechanism for transliteration.
This CL introduces a mechanism into the EnglishVariantsRewriter, which expand variants with prefixed spaces.

While many English words are supported in Mozc, currently there is no way to insert space between words. Therefore, 'ぐーぐるじゃぱん' always becomes 'GoogleJapan',not 'Google Japan'. This is not useful, especially for new words like じぇみにぷろ→'Gemini Pro' and 'ぴくせるふぉーるど'→'Pixel Fold'. By using this mechanism, when the user select 'Pixel' for 'ぴくせる', space-prefixed candidates like ` Fold` are inserted. By selecting that candidate, users can input 'Pixel Fold' as a compound. This will be learned in subsequent inputs.

As there can be several ways to implement the same feature, I guarded this feature with a flag named "english_variation_space_insertion_mode". The current style is mode 1, and we can test mode 2 or later.

PiperOrigin-RevId: 617748976