Add internal support for Gboard Personal Dictionary V1.
This CL introduces support for Gboard Personal Dictionary, which can be exported from Android Gboard.

This is only an internal support and does not include GUI support. But if users select 'auto detection' for the format, Gboard V1 will be automatically detected and imported. We can add support of the GUI format selector in the following CLs.

PiperOrigin-RevId: 619921794