Add a build rule `//win32/installer:executables`
As a preparation to build Mozc installer for Windows (#948), this commit
adds a rule to build all the Mozc executables to be installed with the
correct build configurations (e.g. CPU architecture, CRT link type).

  bazel --bazelrc=windows.bazelrc               \
        build //win32/installer:artifacts       \
        --config oss_windows --cpu=x64_windows

Keep in mind that there still remain many known issues in these
executables. They are not yet ready to correctly run in real devices.

 * Win32 resources are not yet embedded
 * manifest files are not yet embedded
 * some security related compiler / linker flags are not yet set
   (e.g. '/CETCOMPAT' and /DEPENDENTLOADFLAG)

PiperOrigin-RevId: 646329541