Fix macOS builds due to typos in `.bazelrc`
This is a follow up commit to my previous commit [1], which aimed to add

  --noincompatible_enable_cc_toolchain_resolution

build option only to Windows builds as a preparation to support Bazel
build on Windows (#948). It, however, accidentally set the flag to mac
builds, which is why now macOS builds are failing.

With this commit the flag is set only to Windows builds as intended.

#codehealth

 [1]: 2d4eabb3e01c013c85d7c11e41de9b77d3f06186

PiperOrigin-RevId: 646703582