Add platforms and android_binary rule for Android build.
This CL is a preparation for Bzlmod migration.

In Bzlmod, we need to specify the platforms for --android_platforms option.
This CL adds the platforms and android_binary rule for Android build.

This CL also adds a placeholder of rules_android for the compatibility with Bzlmod. This placeholder is necessary because Bzlmod requires rules_android.

https://github.com/google/mozc/issues/1002

PiperOrigin-RevId: 664617381