LSC: Fix null safety issues diagnosed by Clang’s `-Wnonnull` and `-Wnullability`.
These changes are generated by hand.

There are existing callers that call `CreateNewElement()` with a null `next` pointer, and it's pretty obvious that `next` is allowed to be null, to signal the end of the list. It looks as if the previous annotation was in error.

`-Wnonnull` diagnoses cases where a `nullptr` literal is passed to a parameter
annotated nonnull, or where `nullptr` is returned from a function whose return
type is annotated nonnull.

`-Wnullability` diagnoses cases where nullability annotations conflict, for
example between the declaration and definition of a function.

PiperOrigin-RevId: 671779861