Add workarounds to avoid crash issue on macOS 15 (Sequoia)
When a method of imkServer is called, the Mozc process crashes on macOS 15.
This is a workaround to address this crash issue.

As a trade-off of this workaround, events on the candidate window are not handled. (e.g. committing the word by clicking the candidate window does not work).

This workaround should be addressed later.

PiperOrigin-RevId: 675441582