Automated g4 rollback of changelist 678781806.
*** Reason for rollback ***

Potential crash fix.

*** Original change description ***

Mark utf8 functions in Util as deprecated and prefer strings/unicode.h

The following Util functions are marked as deprecated and replaced by the corresponding functions in strings/unicode.h.
- Utf8ToUtf32
- Utf32ToUtf8
- CodepointToUtf8
- CodepointToUtf8Append
- IsValidUtf8

Note that the new encoder handles '\0' as a valid character and characters beyond U+10FFFF as invalid. The new implementation in Unicode standard compliant.

The following functions and classes are deprecated but the curren...
```
***

PiperOrigin-RevId: 679622683