Take `gen_win32_resource_header.py` into account in Bazel build
As a preparation to use Bazel to build Mozc for Windows (#948), this
commit makes sure that gen_win32_resource_header.py is actually used
to generate the final *.rc file to be compiled.

Note that actual Win32 resource compile is still a stub, which will
be implemented in a subsequent commit.

There must be no observable behavior change in GYP build.

PiperOrigin-RevId: 679874821