#1002: Switch to bzlmod for Android build
This is a follow up to our previous commits [1][2], which enabled bzlmod
by default except for Android. With this commit, Android build actually
starts only bzlmod mode.

Basically the idea is to use transition [3] to specify required CPU and
platforms for each libmozc.so target. A bonus is that we no longer need
to rely on rules_android.

 [1]: 1a16c111dffe57479cef376bb29bbb1d03b3377d
 [2]: e7940c16cfab7ee4429270f01e7c5438d2536218
 [3]: https://bazel.build/rules/lib/builtins/transition

PiperOrigin-RevId: 680441153