Use the same fine names for installer when building with Bazel
This is another follow up commit to my previous commits [1], which aimed
to start supporting Mozc installer for Windows with Bazel (#948).

In GYP builds we used to have 64 suffix in the file names of Mozc
installers. Let's do the same thing for now to make our build system
transition simpler.

With this commit 'Mozc64_bazel.msi' finally starts showing up on the
GitHub Actions artifacts section.

 [1]: a60f1f669bf039312365c5abbd241e07b521ffed
 [2]: 615078703f8a9f6e580b323d088301c9da25e398

PiperOrigin-RevId: 680747948