Follow PEP 394 in `build_mozc_for_android.md` (#1194)
This follows up to my previous commit [1], which introduced

  docs/build_mozc_for_android.md

as a dedicated build instruction page for building 'libmozc.so' for
Android as part of removing the dependency on Docker from our build
instructions (#1181).

While the above document uses 'python3' elsewhere, there was one place
where 'python' was used instead. It would be better to avoid 'python'
in favor of PEP 394 [2].

 [1]: b44b056f6ac7db1bd0685eb79823dce49c433b4f
 [2]: https://peps.python.org/pep-0394/

PiperOrigin-RevId: 730872687