Propagate `%ERRORLEVEL%` in the Bazel wrapper (#1220)
This follows up to my previous commit [1], which introduced a Bazel
wrapper batch file.

There was a bug that always reset '%ERRORLEVEL%' to '0' in the wrapper
script, which has prevented any build and/or test failure from being
shown up on the continuous build/test pages. Let's fix it by simply
propagating '%ERRORLEVEL%' in the 'exit' command.

Closes #1219.

 [1]: f0b4074524a4975711cfbf416bbf9c04ddb01838

PiperOrigin-RevId: 737986532